refactor: update auth guard folder structure for better organization and improvements
