--- src/app/features/auth/guards/auth-guard.ts
@@ -1,5 +1,0 @@
-import { CanActivateFn } from '@angular/router';
-
-export const authGuard: CanActivateFn = (route, state) => {
-  return true;
-};
