--- src/app/shared/guards/admin-gurds/auth-guard.ts
@@ -1,22 +1,0 @@
-import { inject } from '@angular/core';
-import { CanActivateFn, Router } from '@angular/router';
-import { map } from 'rxjs';
-import { UserRegisterService } from '../../../features/auth/services/auth.service';
-
-export const authGuard: CanActivateFn = (route, state) => {
-  const _authService = inject(UserRegisterService);
-  const router = inject(Router);
-
-  return _authService.admin$.pipe(
-    map((user) => {
-      const authorized = !!user && user.is_verified;
-      if (!authorized) {
-        console.log('Unauthorized to access');
-        return router.createUrlTree(['/adminlogin']);
-      }
-
-      console.log('Authorized to access');
-      return true;
-    })
-  );
-};
